+++ changelog.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - [Fix unnecessary warnings in native-image build](https://github.com/ballerina-platform/ballerina-standard-library/issues/3861)
+- [Http Compiler Plugin fails for payload type with cyclic fields](https://github.com/ballerina-platform/ballerina-standard-library/issues/3966)
 
 ## [2.5.2] - 2022-12-22